fix(button): 完善button案例
