+++ src/packages/button/demos/h5/demo8.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Button } from '@nutui/nutui-react'
+
+const Demo8 = () => {
+  return (
+    <>
+      <Button block type="primary">
+        Block Button
+      </Button>
+    </>
+  )
+}
+export default Demo8
