+++ src/packages/button/demos/taro/demo2.tsx
@@ -1,0 +1,29 @@
+import React from 'react'
+import { Button } from '@nutui/nutui-react-taro'
+
+const Demo2 = () => {
+  const marginStyle = { margin: 8 }
+  return (
+    <>
+      <Button type="primary" style={marginStyle}>
+        Primary
+      </Button>
+      <Button type="info" style={marginStyle}>
+        Info
+      </Button>
+      <Button type="default" style={marginStyle}>
+        Default
+      </Button>
+      <Button type="danger" style={marginStyle}>
+        Danger
+      </Button>
+      <Button type="warning" style={marginStyle}>
+        Warning
+      </Button>
+      <Button type="success" style={marginStyle}>
+        Success
+      </Button>
+    </>
+  )
+}
+export default Demo2
