+++ src/packages/button/demos/taro/demo5.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import { Star, Plus } from '@nutui/icons-react'
-import { Button } from '@nutui/nutui-react'
+import { Button } from '@nutui/nutui-react-taro'
+import { Star, Plus } from '@nutui/icons-react-taro'
 
-const Demo4 = () => {
+const Demo5 = () => {
   const marginStyle = { margin: 8 }
   return (
     <>
@@ -72,7 +72,14 @@ const Demo4 = () => {
       >
         Disabled
       </Button>
-      <Button disabled icon={<Star />} rightIcon={<Star />} style={marginStyle}>
+      <Button
+        disabled
+        icon={<Star />}
+        rightIcon={<Star />}
+        style={{
+          margin: 8,
+        }}
+      >
         Disabled
       </Button>
       <Button
@@ -80,28 +87,29 @@ const Demo4 = () => {
         shape="square"
         fill="outline"
         type="primary"
-        icon={<Plus width="20" />}
+        icon={<Plus size="20" />}
         style={marginStyle}
       />
       <Button
         disabled
         type="primary"
-        icon={<Plus width="20" />}
+        icon={<Plus size="20" />}
         style={marginStyle}
       />
       <Button
         disabled
         type="primary"
         fill="dashed"
-        icon={<Plus width="20" />}
+        icon={<Plus size="20" />}
         style={marginStyle}
       />
       <Button
         disabled
+        shape="round"
         type="primary"
         size="large"
-        icon={<Star width={20} height={20} />}
-        rightIcon={<Star width={20} height={20} />}
+        icon={<Star size={20} />}
+        rightIcon={<Star size={20} />}
         style={marginStyle}
       >
         Disabled
@@ -109,4 +117,4 @@ const Demo4 = () => {
     </>
   )
 }
-export default Demo4
+export default Demo5
