+++ src/packages/button/demos/taro/demo7.tsx
@@ -1,0 +1,28 @@
+import React, { useState } from 'react'
+import { Button } from '@nutui/nutui-react-taro'
+
+const Demo7 = () => {
+  const [loading, setLoading] = useState(false)
+  const marginStyle = { margin: 8 }
+  return (
+    <>
+      <Button loading type="warning" style={marginStyle}>
+        Loading
+      </Button>
+      <Button
+        loading={loading}
+        type="success"
+        onClick={() => {
+          setTimeout(() => {
+            setLoading(false)
+          }, 1500)
+          setLoading(!loading)
+        }}
+        style={marginStyle}
+      >
+        Click me!
+      </Button>
+    </>
+  )
+}
+export default Demo7
