+++ src/packages/button/demos/taro/demo9.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Button } from '@nutui/nutui-react-taro'
+
+const App = () => {
+  return (
+    <>
+      <Button block type="primary">
+        Block Button
+      </Button>
+    </>
+  )
+}
+export default App
