+++ src/packages/button/doc.md
@@ -58,23 +58,7 @@ import { Button } from '@nutui/nutui-react';
 
 :::demo
 
-```tsx
-import React from "react";
-import { Button } from '@nutui/nutui-react';
-
-const App = () => {
-  const marginStyle = { margin: 8 }
-  return (
-    <>
-      <Button style={marginStyle} type="primary" shape="square">
-        方形按钮
-      </Button>
-      <Button style={marginStyle} type="primary">圆形按钮</Button>
-    </>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo5.tsx'></CodeBlock>
 
 :::
 
@@ -82,37 +66,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React, { useState } from "react";
-import { Button } from '@nutui/nutui-react';
-
-const App = () => {
-  const [loading, setLoading] = useState(false)
-  const marginStyle = { margin: 8 }
-  return (
-    <>
-      <Button loading type="warning" style={marginStyle}>
-        加载中
-      </Button>
-      <Button
-        loading={loading}
-        type="success"
-        onClick={() => {
-          setTimeout(() => {
-            setLoading(false)
-          }, 1500)
-          setLoading(!loading)
-        }}
-        style={marginStyle}
-      >
-        Click me!
-      </Button>
-    </>
-  );
-};
-export default App;
-```
-
+<CodeBlock src='h5/demo6.tsx'></CodeBlock>
 :::
 
 ### 按钮尺寸
@@ -121,28 +75,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Button } from '@nutui/nutui-react';
-
-const App = () => {
-  const marginStyle = { margin: 8 }
-  return (
-    <>
-      <Button size="large" type="primary">大号按钮</Button>
-      <Button type="primary" style={marginStyle}>普通按钮</Button>
-      <Button style={marginStyle}>普通按钮</Button>
-      <Button size="small" style={marginStyle} type="primary">
-        小型按钮
-      </Button>
-      <Button size="mini" style={marginStyle} type="primary">
-        迷你按钮
-      </Button>
-    </>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo7.tsx'></CodeBlock>
 
 :::
 
@@ -152,19 +85,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Button } from '@nutui/nutui-react';
-
-const App = () => {
-  return (
-    <>
-      <Button block type="primary">块级元素</Button>
-    </>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo8.tsx'></CodeBlock>
 
 :::
 
@@ -174,62 +95,7 @@ export default App;
 
 :::demo
 
-```tsx
-import React from "react";
-import { Button } from '@nutui/nutui-react';
-
-const App = () => {
-  const marginStyle = { margin: 8 }
-  return (
-    <>
-      <Button
-        style={{
-          margin: 8,
-          '--nutui-button-default-border-color': 'blue',
-          '--nutui-button-default-color': '#fff',
-          '--nutui-button-default-background-color': 'blue',
-        }}
-      >
-        单色按钮
-      </Button>
-      <Button
-        fill="outline"
-        style={{
-          margin: 8,
-          '--nutui-button-default-border-color': '#7232dd',
-          '--nutui-button-default-color': '#7232dd',
-        }}
-      >
-        单色按钮
-      </Button>
-      <Button
-        style={{
-          margin: 8,
-          '--nutui-button-default-border-color': 'transparent',
-          '--nutui-button-default-color': '#fff',
-          '--nutui-button-default-background-color':
-            'rgba(10,101,208,0.75)',
-        }}
-      >
-        单色按钮
-      </Button>
-      <Button
-        type="primary"
-        style={{
-          margin: 8,
-          '--nutui-button-default-border-color': 'transparent',
-          '--nutui-button-default-color': '#fff',
-          '--nutui-button-default-background-color':
-            'linear-gradient(to right, #ff6034, #ee0a24)',
-        }}
-      >
-        渐变按钮
-      </Button>
-    </>
-  );
-};
-export default App;
-```
+<CodeBlock src='h5/demo9.tsx'></CodeBlock>
 
 :::
